ser-model/requirements.txt
numpy
Pillow
--extra-index-url https://google-coral.github.io/py-repo/ 
tflite_runtime
